Refactored

+++ src/main/java/com/capgemini/boardgames/dto/GameDto.java
@@ -9,13 +9,13 @@ import java.util.List;
 @Setter
 public class GameDto {
 
-    private int id;
+    private long id;
     private String name;
     private int minPlayerNumber;
     private int maxPlayerNumber;
     private List<String> subscribersList;
 
-    public GameDto(int id, String name, int minPlayerNumber, int maxPlayerNumber) {
+    public GameDto(long id, String name, int minPlayerNumber, int maxPlayerNumber) {
         this.id = id;
         this.name = name;
         this.minPlayerNumber = minPlayerNumber;
